trans: Move TransItem to its own module.
